Finished removing all references to Peano curves.  The working code is on my desktop.  The code can be merged in at a later time.

+++ AnnotationDisplay.cpp
@@ -168,7 +168,7 @@ string AnnotationDisplay::mouseClick(point2D pt)
 
 int AnnotationDisplay::width()
 {
-	return (int)max_width*2;
+    return max_width*2;
 }
